pause: working on userContext
