+++ src/App.jsx
@@ -1,6 +1,7 @@
 import { AppRoutes } from 'routes';
 import { useState } from 'react';
-import { getUserFromLS, UserContext } from 'store/context';
+import { UserContext } from 'store/context';
+import { getUserFromLS } from 'helpers/helpers';
 import { GlobalStyle } from 'components';
 
 function App() {
@@ -8,10 +9,12 @@ function App() {
 
   const isAllowed = Boolean(currentUser);
   return (
-    <UserContext.Provider value={{ currentUser, setCurrentUser }}>
+    // <UserContext.Provider value={{ currentUser, setCurrentUser }}>
+    <>
       <GlobalStyle />
       <AppRoutes isAllowed={isAllowed} />
-    </UserContext.Provider>
+    </>
+    // </UserContext.Provider>
   );
 }
