+++ src/hooks/useUser.jsx
@@ -1,0 +1,4 @@
+import { useContext } from "react";
+import { UserContext } from "providers/UserContextProvider";
+
+export const useUser = () => useContext(UserContext);
